fix(try): implement data fetching combining Suspense and Use
