+++ packages/try/src/Use/index.story.tsx
@@ -1,7 +1,5 @@
-import { css } from '@emotion/css';
-import { gutter } from '@learn-react/core/src/helpers/Style';
-import { type ChangeEvent, Suspense, use, useState } from 'react';
-import { requestGetUser } from './api';
+import { Suspense, use } from 'react';
+import { requestGetUsers } from './api';
 import { type User } from './model';
 
 /**
@@ -12,24 +10,12 @@ import { type User } from './model';
  * @remarks
  * `use` の引数に渡す Promise は原則としてサーバーコンポーネント内で生成することが推奨されます。
  * クライアントコンポーネント内で生成するとレンダリングの度に新しい Promise が生成され、無限ループに陥ります。
- * これを回避するには、 `useState` の初期値に関数を渡すことで初回のみ実行されるといったハックが必要です。
+ * これを回避するには、 `use` を使うコンポーネントの外で Promise を生成し、それを Props として渡すようにします。
  *
  * @see {@link https://ja.react.dev/reference/react/use use}
  */
 export const Story = () => {
-  const [dataPromise, setDataPromise] = useState(() =>
-    requestGetUser({
-      path: '1',
-    }),
-  );
-
-  const handleChange = ({ target: { value } }: ChangeEvent<HTMLSelectElement>) => {
-    setDataPromise(
-      requestGetUser({
-        path: value,
-      }),
-    );
-  };
+  const dataPromise = requestGetUsers();
 
   return (
     <>
@@ -37,37 +23,20 @@ export const Story = () => {
 
       <p>React 組み込み API である Suspense と Use を組み合わせたデータ取得デモ。</p>
 
-      <label className={styleSelectForm}>
-        <span>Select User: </span>
-        <select onChange={handleChange}>
-          {[...Array(10).keys()].map((index) => (
-            <option key={index} value={index + 1}>
-              User {index + 1}
-            </option>
-          ))}
-        </select>
-      </label>
-
       <hr />
 
       <Suspense fallback={<div>Loading...</div>}>
-        <User dataPromise={dataPromise} />
+        <Users dataPromise={dataPromise} />
       </Suspense>
     </>
   );
 };
 
-const styleSelectForm = css`
-  display: flex;
-  gap: ${gutter(2)};
-  align-items: center;
-`;
-
 type InternalProps = {
-  dataPromise: Promise<User>;
+  dataPromise: Promise<User[]>;
 };
 
-const User = ({ dataPromise }: InternalProps) => {
+const Users = ({ dataPromise }: InternalProps) => {
   const data = use(dataPromise);
 
   return (
